Remove old sanitizeNumber
