+++ src/main.ts
@@ -37,10 +37,9 @@ import {
 	productionLoggerMiddleware,
 } from './middleware/logger_middleware';
 import { validateAddressMiddleware } from './middleware/validations_middleware';
-import { sanitizeData } from './sanitize';
 import { TxRequest, TxRequestBody } from './types/request_types';
-import { parseBlockNumber } from './utils';
-// import { sanitizeNumbers } from './utils';
+import { sanitizeNumbers } from './utils/sanitize';
+import { parseBlockNumber } from './utils/utils';
 
 async function main() {
 	const configOrNull = Config.GetConfig();
@@ -82,9 +81,7 @@ async function main() {
 	) {
 		app.get(path, async (req, res, next) => {
 			try {
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-				// res.send(sanitizeNumbers(await cb(req.params)));
-				res.send(sanitizeData(await cb(req.params)));
+				res.send(sanitizeNumbers(await cb(req.params)));
 			} catch (err) {
 				return next(err);
 			}
@@ -102,8 +99,7 @@ async function main() {
 	) {
 		app.post(path, async (req: TxRequest, res, next) => {
 			try {
-				// res.send(sanitizeNumbers(await cb(req.params, req.body)));
-				res.send(sanitizeData(await cb(req.params, req.body)));
+				res.send(sanitizeNumbers(await cb(req.params, req.body)));
 			} catch (err) {
 				return next(err);
 			}
